requirements.txt
# requirements.txt
fastmcp
proxmoxer
python-dotenv
